requirements.txt
# YOLOv8 绝缘子破损检测系统依赖包

# 深度学习框架
torch>=2.0.0
torchvision>=0.15.0

# YOLOv8
ultralytics>=8.0.0

# 数据处理
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=10.0.0
pandas>=2.0.0

# 可视化
matplotlib>=3.7.0
seaborn>=0.12.0

# 工具库
tqdm>=4.65.0
pyyaml>=6.0

# 可选：GPU 支持（需要根据 CUDA 版本安装）
# 例如：torch==2.0.0+cu118 -f https://download.pytorch.org/whl/torch_stable.html
